requirements.txt
deepagents==0.0.5
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.76
openai==1.108.0
python-dotenv==1.0.0
rich==14.1.0
typing-extensions==4.15.0
pydantic>=2.11.9
jupyter>=1.0.0
ipywidgets>=8.1.1
pandas>=2.1.4
matplotlib>=3.8.2
ddgs>=9.6.0
